+++ nevergrad/optimization/recorded_recommendations.csv
@@ -88,6 +88,7 @@ FastGADiscreteOnePlusOne,0.7531428339,1.095956118,0.0,1.3423563714,,,,,,,,,,,,
 FastGANoisyDiscreteOnePlusOne,0.7531428339,1.095956118,0.0,1.3423563714,,,,,,,,,,,,
 FastGAOptimisticNoisyDiscreteOnePlusOne,0.7531428339,1.095956118,0.0,1.3423563714,,,,,,,,,,,,
 GeneticDE,1.012515477,-0.9138691467,-1.0295302074,1.2097964496,,,,,,,,,,,,
+HSDE,0.5,-0.7999999785,-3.3e-09,4.0000000001,5.0000000231,2.7015115302,-2.080734155,-4.9499624832,,,,,,,,
 HaltonSearch,-0.318639364,-0.7647096738,-0.7063025628,1.0675705239,,,,,,,,,,,,
 HaltonSearchPlusMiddlePoint,0.0,0.0,0.0,0.0,,,,,,,,,,,,
 HammersleySearch,0.2104283942,-1.1503493804,-0.1397102989,0.8416212336,,,,,,,,,,,,
@@ -109,6 +110,7 @@ IsoEMNATBPSA,0.0,0.0,0.0,0.0,,,,,,,,,,,,
 LHSSearch,-0.3978418928,0.827925915,1.2070034191,1.3637174061,,,,,,,,,,,,
 LargeHaltonSearch,-67.4489750196,43.0727299295,-25.3347103136,-56.5948821933,,,,,,,,,,,,
 LhsDE,-0.8072358182,0.6354687554,1.575403308,1.1808277036,2.5888168575,-0.1627990771,-3.656466139,-1.040475202,,,,,,,,
+LhsHSDE,-0.8072358182,0.6354687554,1.575403308,1.1808277036,2.5888168575,-0.1627990771,-3.656466139,-1.040475202,,,,,,,,
 MetaCauchyRecentering,1.8789278226,-0.2085387973,-1.3832372686,3.9852740423,,,,,,,,,,,,
 MetaModel,1.012515477,-0.9138805701,-1.029555946,1.2098418178,,,,,,,,,,,,
 MetaModelDiagonalCMA,1.012515477,-0.9138691467,-1.0295302074,1.2097964496,,,,,,,,,,,,
